2025-02-27/store/pyproject.toml
[tool.poetry]
name = "store"
version = "0.1.0"
description = ""
authors = ["OKnischay <nischaybasi015@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
pygments = "^2.19.1"
django-environ = "^0.12.0"
django-environs = "^0.1.1"
psycopg = "^3.2.5"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
